refactor: controller 를 CarRacingMain으로 변경
