--- src/main/java/Application.java
@@ -1,9 +1,0 @@
-import controller.RacingCarGame;
-import view.InputView;
-
-public class Application {
-    public static void main(String[] args) {
-        RacingCarGame racingCarGame = new RacingCarGame();
-        racingCarGame.run();
-    }
-}
